Introduce EncryptorBuilder to make JCE optional
The default encryptors so far all use a default keyLength of 256, and
doing so is only possible when running in a JRE where the JCE is installed.

At the same time, the Encryptors factory methods are not flexible enough
to account for easy configuration of this. Therefore, introduce the
EncryptorBuilder as a starting point for enveloping the Encryptors with
a more user-friendly interface.

Move CipherAlgorithm to be its own public class as it is exposed for the
EncryptorBuilder.

On the test-side, make the encryption tests take the JCE availability
into account, so 128 is used as keyLength if needed.

Fixes gh-2917